remove svelte files, rename protocol lib
