+++ src/index.js
@@ -1,2 +1,2 @@
-export { nftApi } from './lib/me3-protocol.js'
+export { claim, nftApi } from './lib/protocol.js'
 export * from './me3-claim-widget.js'
