--- src/lib/Input.svelte
@@ -1,72 +1,0 @@
-<script>
-  export let label
-  export let placeholder
-  export let value
-  export let domain
-
-  export let highlightError = false
-  export let validations = []
-  export let hasError = !value
-
-  let field
-  let id = `field-${Date.now()}`
-
-  function validate (ev) {
-    const { value } = ev.target
-    hasError = !(validations.every(fn => fn(value)))
-
-    if (highlightError && !hasError) highlightError = false
-  }
-</script>
-
-<label for={id}>{label}</label>
-
-<div class="fake-input" class:error={highlightError}>
-  <input
-    bind:this={field}
-    on:input={validate}
-    bind:value={value}
-    type="text"
-    placeholder={placeholder}
-    name={id}
-    {id}
-    />
-    <span>.{domain}</span>
-</div>
-
-<style>
-  .error {
-    border: 1px solid red;
-  }
-
-  label {
-    display: var(--me3-label-display, block);
-    margin: var(--me3-label-margin, 0 0 8px 0);
-    font-size: var(--me3-label-font-size, 14px);
-    line-height: var(--me3-label-line-height, 24px);
-    font-weight: var(--me3-label-font-weight, 500);
-  }
-
-  input {
-    border: 0;
-    text-align: right;
-    outline: none;
-    width: 100%;
-    background: transparent;
-    color: var(--me3-input-text-color, #1c1c33);
-    margin: 0;
-    padding: 0;
-  }
-
-  .fake-input {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    border-radius: var(--me3-input-border-radius, 16px);
-    border: var(--me3-input-border, 1px solid #dedede);
-    padding: var(--me3-input-padding, 16px);
-    margin: var(--me3-input-margin, 0);
-    background: var(--me3-input-background, #ffffff);
-  }
-</style>
